+++ zeyvro_turnstile.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * Zeyvro - Cloudflare Turnstile for PrestaShop
  *
@@ -47,7 +48,7 @@ class Zeyvro_Turnstile extends Module
     {
         $this->name = 'zeyvro_turnstile';
         $this->tab = 'front_office_features';
-        $this->version = '1.1.1';
+        $this->version = '1.1.2';
         $this->author = 'Zeyvro';
         $this->need_instance = 0;
         $this->ps_versions_compliancy = ['min' => '8.0.0', 'max' => '9.99.99'];